backend/requirements.txt
Django==5.0
djangorestframework
psycopg2-binary
google-generativeai
django-cors-headers
